Moving the audit trail up a level
