+++ server/domain/application.js
@@ -1,6 +1,9 @@
 'use strict';
 
 var config = require('config');
+var audit = require('./audit');
+var acl = require('./acl');
+
 var application = function() {
   switch (config.dataSource.toLowerCase()) {
     case 'etcd':
@@ -11,13 +14,41 @@ var application = function() {
   }
 };
 
+var getUserDetails = function (req) {
+    return config.RequiresAuth ? req.user._json : {name: 'Anonymous'};
+};
+
 module.exports = {
     getApplications: function (cb) {
       application().getApplications(cb);
     },
 
     addApplication: function (applicationName, req, cb) {
-      application().addApplication(applicationName, req, cb);
+      application().addApplication(applicationName, req, function(err) {
+        if (err) {
+            return cb(err);
+        }
+
+        audit.addApplicationAudit(getUserDetails(req), applicationName, 'Created', function () {
+            if (err) {
+                cb(err);
+                return;
+            }
+        });
+
+        if (config.RequiresAuth) {
+            var userEmail = getUserDetails(req).email.toLowerCase(); // todo: need better user management
+            acl.grant(userEmail, applicationName, function (grantErr) {
+                if (grantErr) {
+                    cb(grantErr);
+                    return;
+                }
+                cb();
+            });
+        } else {
+            cb();
+        }
+      });
     },
 
     deleteApplication: function (applicationName, req, cb) {
